+++ jest.config.js
@@ -70,12 +70,12 @@ module.exports = {
   // ],
 
   // An array of file extensions your modules use
-  // moduleFileExtensions: [
-  //   "js",
-  //   "json",
-  //   "jsx",
-  //   "node"
-  // ],
+  moduleFileExtensions: [
+    "js",
+    "json",
+    "jsx",
+    "node"
+  ],
 
   // A map from regular expressions to module names that allow to stub out resources with a single module
   // moduleNameMapper: {},
@@ -126,7 +126,7 @@ module.exports = {
 
   // The path to a module that runs some code to configure or set up the testing framework before each test
   // -> Great place to run complex code that creates required dependencies, eg: i18n instances
-  // setupTestFrameworkScriptFile: null,
+  setupTestFrameworkScriptFile: "jest-extended",
 
   // A list of paths to snapshot serializer modules Jest should use for snapshot testing
   // snapshotSerializers: [],